profile/requirements.txt
blinker==1.9.0
click==8.3.0
Flask==3.1.2
Flask-DebugToolbar==0.14.1
Flask-SQLAlchemy==3.1.1
flask-thumbnails==1.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
pillow==12.0.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
